requirements.txt
pandas==1.5.2
tqdm @ file:///tmp/build/80754af9/tqdm_1602185206534/work
python-binance==1.0.16
numpy==1.23.4
DateTime==4.3
discord==2.3.2
discord.py==2.3.2
requests==2.29.0
regex @ file:///C:/ci/regex_1602770567298/work
urllib3==1.26.7
flipside==2.0.8
